// src/types/sectionOne/types.ts
export interface ItemsTypes {
  city?: string;
  date?: string;
  weatherimg?: string;
  temperature?: string;
  bgcolor?: string;
}

export interface WeatherTypes {
  src: string;
  alt: string;
  name: string;
}

export interface InputDatasTypes {
  inputName: string;
  stateName: string;
  defaultValue: string;
}